Update fn-ctrl so that it ignores user set annotations on fn's deployment
